// backend/src/config/env.d.ts
export declare const env: {
    PORT: string;
    NODE_ENV: string;
    MONGODB_URI: string;
    ACCESS_TOKEN_SECRET: string;
    REFRESH_TOKEN_SECRET: string;
    GOOGLE_CLIENT_ID: string;
    GOOGLE_CLIENT_SECRET: string;
    GOOGLE_CALLBACK_URL: string;
    SMTP_HOST: string;
    SMTP_PORT: number;
    SMTP_SECURE: boolean;
    SMTP_USER: string;
    SMTP_PASS: string;
    MAIL_FROM: string;
    FRONTEND_URL: string;
};
//# sourceMappingURL=env.d.ts.map